reorganize modules to better match dependencies.
