+++ src/models/combat/ai_impl.py
@@ -7,8 +7,8 @@ from typing import Callable, Sequence, Set
 from models.characters.character_base import Character
 from models.characters.states import Attributes
 from models.characters.subroutines_base import build_subroutine
-from models.combat.ai_base import AI
-from models.combat.moves_base import Move
+from models.characters.ai_base import AI
+from models.characters.moves_base import Move
 
 SelectionFun = Callable[[Sequence[Move]], Move]
