+++ src/models/combat/combat_logic.py
@@ -5,7 +5,7 @@ from events.events_base import EventListener, EventType
 from models.characters.character_base import Character
 from models.characters.states import Attributes
 from models.combat.combat_stack import CombatStack
-from models.combat.moves_base import Move
+from models.characters.moves_base import Move
 
 
 class CombatLogic(EventListener):
