+++ src/simulation/combat_manager_base.py
@@ -4,7 +4,7 @@ from typing import List, Sequence, Tuple
 from events.events_base import EventManager, MoveExecutedEvent
 from models.characters.character_base import Character
 from models.characters.conditions import IsDead
-from models.combat.moves_base import Move
+from models.characters.moves_base import Move
 
 GroupMove = Sequence[Move]
 GroupMoveSet = Sequence[GroupMove]
