+++ src/views/artists/combat_stack_artist.py
@@ -1,10 +1,10 @@
 import logging
-from typing import Dict, List, Optional, Tuple
+from typing import Dict, List, Optional
 
 from pygame.rect import Rect
 
 from data.colors import DARK_GRAY, LIGHT_GRAY, RED, WHITE, YELLOW
-from models.combat.moves_base import Move
+from models.characters.moves_base import Move
 from models.scenes.combat_scene import CombatScene, MoveData
 from models.scenes.scenes_base import Scene
 from views.artists.drawing_utils import rescale_horizontal
